requirements.txt
accelerate==1.2.1
albucore==0.0.16
albumentations==1.4.15
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
comfy==0.0.1
diffusers==0.32.1
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.16.1
flet==0.25.2
fsspec==2024.12.0
gdown==5.2.0
gguf==0.14.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
inquirerpy==0.3.4
Jinja2==3.1.5
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.1
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optimum-quanto==0.2.6
packaging==24.2
peft==0.14.0
pfzy==0.3.4
pillow==11.1.0
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.5
pydantic_core==2.27.2
PySocks==1.7.1
PyYAML==6.0.2
regex==2024.11.6
repath==0.9.0
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.0
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
termcolor==2.5.0
tifffile==2025.1.10
timm==1.0.13
tokenizer==3.4.5
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.0
transparent-background==1.3.3
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wget==3.2
zipp==3.21.0
